+++ templates/master-schedule-list.php
@@ -76,6 +76,10 @@ foreach ( $weekdays as $weekday ) {
 
 	if ( !$skip_day ) {
 
+		// 2.3.3.6: set next and previous day for split shift IDs
+		$nextday = radio_station_get_next_day( $weekday );
+		$prevday = radio_station_get_previous_day( $weekday );
+
 		// 2.3.2: move up time calculations for optional date display
 		$day_start_time = radio_station_to_time( $weekdates[$weekday] . ' 00:00' );
 		$day_end_time = $day_start_time + ( 24 * 60 * 60 );
@@ -150,6 +154,7 @@ foreach ( $weekdays as $weekday ) {
 			foreach ( $shifts as $shift ) {
 
 				$show = $shift['show'];
+				$split_id = false;
 
 				// --- convert shift time data ---
 				// 2.3.2: replace strtotime with to_time for timezones
@@ -182,9 +187,11 @@ foreach ( $weekdays as $weekday ) {
 					if ( isset( $shift['real_start'] ) ) {
 						$real_shift_start = radio_station_convert_shift_time( $shift['real_start'] );
 						$real_shift_start = radio_station_to_time( $weekdate . ' ' . $real_shift_start ) - ( 24 * 60 * 60 );
+						$split_id = strtolower( $prevday . '-' . $weekday );
 					} elseif ( isset( $shift['real_end'] ) ) {
 						$real_shift_end = radio_station_convert_shift_time( $shift['real_end'] );
 						$real_shift_end = radio_station_to_time( $weekdate . ' ' . $real_shift_end ) + ( 24 * 60 * 60 );
+						$split_id = strtolower( $weekday . '-' . $nextday );
 					}
 				}
 
@@ -194,6 +201,7 @@ foreach ( $weekdays as $weekday ) {
 					$show_link = apply_filters( 'radio_station_schedule_show_link', $show['url'], $show['id'], 'list' );
 				}
 
+				// --- list item classes ---
 				// 2.3.0: add genre classes for highlighting
 				$classes = array( 'master-list-day-item' );
 				$terms = wp_get_post_terms( $show['id'], RADIO_STATION_GENRES_SLUG, array() );
@@ -205,6 +213,11 @@ foreach ( $weekdays as $weekday ) {
 				// 2.3.2: check for now playing shift
 				if ( ( $now >= $shift_start_time ) && ( $now < $shift_end_time ) ) {
 					$classes[] = 'nowplaying';
+				}
+				// 2.3.3.6: add overnight split ID for highlighting
+				if ( $split_id ) {
+					$classes[] = 'overnight';
+					$classes[] = 'split-' . $split_id;
 				}
 
 				// --- open show list item ---